Use function for creating server instead of global
In an upcoming commit, I need to add a custom middleware only during unit testing.  We now can instantiate separate instances of the app, to prevent tests from polluting each other's environment.
